fix: fix minor bugs
